# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API бэкенда видеоконференций.",
        "title": "Video Conference API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/meetings/reminders/status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meetings"
                ],
                "summary": "Статус подписки на напоминание",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID комнаты",
                        "name": "room_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.ReminderStatusSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/meetings/reminders/subscribe": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meetings"
                ],
                "summary": "Подписаться на напоминание",
                "parameters": [
                    {
                        "description": "room_id и email",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.SubscribeReminderBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Уже подписан",
                        "schema": {
                            "$ref": "#/definitions/controllers.SimpleMessageSuccess"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.SubscribeReminderCreated"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.ErrorJSON": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Неверные данные запроса"
                }
            }
        },
...
        "controllers.ReminderStatusSuccess": {
            "type": "object",
            "properties": {
                "subscribed": {
                    "type": "boolean"
                }
            }
        },
        "controllers.SimpleMessageSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.SubscribeReminderBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "room_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "controllers.SubscribeReminderCreated": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}